init2

+++ resources/js/components/incomes/List.vue
@@ -20,37 +20,38 @@
                     Buscar
                 </button>
             </div>
-            <table class="table">
-                <thead>
+            <VTable :data="income" class="table table-dark mt-3">
+                <template #head>
                     <tr>
-                        <th scope="col">Producto</th>
-                        <th scope="col">Cantidad</th>
-                        <th scope="col">Fecha</th>
-                        <th scope="col">Registro</th>
-                        <th scope="col"></th>
+                        <VTh sortKey="product">Producto</VTh>
+                        <th>Cantidad</th>
+                        <th>Usuario</th>
+                        <th>Fecha</th>
+                        <th>Registro</th>
+                        <th></th>
                     </tr>
-                </thead>
-                <tbody>
-                    <tr v-for="item in income" :key="item.id">
-                        <td>{{ item.product }}</td>
-                        <td>{{ item.quantity }}</td>
-                        <td>{{ item.user }}</td>
-                        <td>{{ item.date_income }}</td>
-                        <td>{{ item.created_at }}</td>
+                </template>
+                <template #body="{ rows }">
+                    <tr v-for="row in rows" :key="row.id">
+                        <td>{{ row.product }}</td>
+                        <td>{{ row.quantity }}</td>
+                        <td>{{ row.user }}</td>
+                        <td>{{ row.date_income }}</td>
+                        <td>{{ row.created_at }}</td>
 
                         <td>
                             <button
                                 v-can="'eliminar entrada'"
                                 type="button"
-                                @click="destroy(item.id)"
+                                @click="destroy(row.id)"
                                 class="btn bg-danger btn-sm"
                             >
                                 <i class="fi fi-trash"></i>
                             </button>
                         </td>
                     </tr>
-                </tbody>
-            </table>
+                </template>
+            </VTable>
         </div>
 
         <div class="alert alert-primary" role="alert">
@@ -144,6 +145,7 @@ export default {
     methods: {
         getList() {
             this.$store.dispatch("Incomeactions", date_now);
+            this.getDate();
         },
         getList_two() {
             this.$store.dispatch("Incomeactions", this.dateList);
@@ -157,17 +159,25 @@ export default {
         },
 
         destroy(id) {
-            let url = this.urlincome + "/" + id;
-            let response = axios.delete(url);
-            try {
-                this.$store.dispatch("Incomeactions", date_now);
-                Swal.fire({
-                    title: `${response.data.message}`,
-                    icon: "success",
-                });
-            } catch (error) {
-                console.log(error);
-            }
+            Swal.fire({
+                title: "Deseas eliminar la entrada?",
+                showCancelButton: true,
+                confirmButtonText: "Si",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    let url = this.urlincome + "/" + id;
+                    let response = axios.delete(url);
+                    try {
+                        this.getList();
+                        Swal.fire({
+                            title: `${response.data.message}`,
+                            icon: "success",
+                        });
+                    } catch (error) {
+                        console.log(error);
+                    }
+                }
+            });
         },
     },
 };
